Restructure orderModel
Change all client reference to orderModel.
