+++ README.md
@@ -1,5 +1,5 @@
 # Fast-Food-Fast
-[![Build Status](https://travis-ci.com/Bobsar0/Fast-Food-Fast.svg?branch=developer-ui)](https://travis-ci.com/Bobsar0/Fast-Food-Fast)
+[![Build Status](https://travis-ci.com/Bobsar0/Fast-Food-Fast.svg?branch=Bobsar0-patch-1)](https://travis-ci.com/Bobsar0/Fast-Food-Fast)
 [![Coverage Status](https://coveralls.io/repos/github/Bobsar0/Fast-Food-Fast/badge.svg)](https://coveralls.io/github/Bobsar0/Fast-Food-Fast)
 [![Code Climate](https://codeclimate.com/github/codeclimate/codeclimate/badges/gpa.svg)](https://codeclimate.com/github/Bobsar0/Fast-Food-Fast)
