+++ api/v1/controllers/ordersController.js
@@ -1,32 +1,10 @@
 import _ from 'lodash';
 import BaseController from './baseController';
-// Import test client that works with JS data structures
-// import client from '../client';
-
-// // API test store
-// const store = [{
-//   orderId: 'ABCDEFGHIJKLMNO',
-//   name: 'Chicken',
-//   price: 1000,
-//   quantity: 2,
-//   username: 'Steve',
-//   userAddr: 'Andela',
-//   userRank: 'admin',
-// },
-// {
-//   orderId: 'PQRSTUVWXYZABCDE',
-//   name: 'Meatpie',
-//   price: 750,
-//   quantity: 3,
-//   username: 'Anonymous',
-//   userAddr: 'Somewhere',
-//   userRank: 'guest',
-// }];
 
 export default class extends BaseController {
   constructor(client, username) {
     super(client, username);
-    this.client = client;
+    this.orders = client;
   }
 
   // GET /orders/:orderId
@@ -34,10 +12,10 @@ export default class extends BaseController {
   // Error is caught on server
   read(id = '') {
     if (id === '') {
-      return this.client.getAll()
+      return this.orders.getAll()
         .then(store => _.map(store, order => _.merge({ orderId: order.orderId }, order)));
     }
-    return this.client.get(id)
+    return this.orders.get(id)
       .then(res => new Promise((resolve, reject) => {
         if (res.found) {
           return resolve(res.value);
@@ -50,7 +28,7 @@ export default class extends BaseController {
   // POST /orders
   // create makes a new order and returns orderId merged with order if successful
   create(order) {
-    return this.client.save({
+    return this.orders.save({
       body: order,
     }) // Merge the results obtained from save() which will be rendered by server
       .then(res => _.merge(res, { orderId: res.orderId, order: res.order }))
@@ -60,7 +38,7 @@ export default class extends BaseController {
   // PUT /orders/:orderId
   // update modifies an order and returns orderId if successful
   update(id, attrs) {
-    return this.client.update({
+    return this.orders.update({
       orderId: id,
       body: attrs,
     })
@@ -75,7 +53,7 @@ export default class extends BaseController {
 
   // DELETE /orders/:id
   delete(id) {
-    return this.client.cancel(id)
+    return this.orders.cancel(id)
       .then(res => new Promise((resolve, reject) => {
         if (res.deleted) {
           return resolve(res);
